+++ packages/tenancy/src/application/index.ts
@@ -68,3 +68,24 @@ export type {
   AcceptInvitationDeps,
 } from "./accept-invitation.js";
 export { acceptInvitation } from "./accept-invitation.js";
+
+export type { OrganizationSummary, GetOrganizationDeps } from "./get-organization-query.js";
+export { toOrganizationSummary, getOrganization } from "./get-organization-query.js";
+
+export type {
+  OrganizationMemberSummary,
+  ListOrganizationMembersDeps,
+} from "./list-organization-members-query.js";
+export {
+  toOrganizationMemberSummary,
+  listOrganizationMembers,
+} from "./list-organization-members-query.js";
+
+export type {
+  PendingInvitationSummary,
+  ListPendingInvitationsDeps,
+} from "./list-pending-invitations-query.js";
+export {
+  toPendingInvitationSummary,
+  listPendingInvitations,
+} from "./list-pending-invitations-query.js";