+++ packages/tenancy/src/index.ts
@@ -136,3 +136,30 @@ export type {
   AcceptInvitationDeps,
 } from "./application/accept-invitation.js";
 export { acceptInvitation } from "./application/accept-invitation.js";
+
+export type {
+  OrganizationSummary,
+  GetOrganizationDeps,
+} from "./application/get-organization-query.js";
+export {
+  toOrganizationSummary,
+  getOrganization,
+} from "./application/get-organization-query.js";
+
+export type {
+  OrganizationMemberSummary,
+  ListOrganizationMembersDeps,
+} from "./application/list-organization-members-query.js";
+export {
+  toOrganizationMemberSummary,
+  listOrganizationMembers,
+} from "./application/list-organization-members-query.js";
+
+export type {
+  PendingInvitationSummary,
+  ListPendingInvitationsDeps,
+} from "./application/list-pending-invitations-query.js";
+export {
+  toPendingInvitationSummary,
+  listPendingInvitations,
+} from "./application/list-pending-invitations-query.js";